Revert "Flow UI: Long string out of bounds fix (#2683)"
This reverts commit 14b9e861e858b8a6a1e7d285b1a94664b0abf741.
