+++ web/ui/src/features/component/ComponentList.module.css
@@ -22,3 +22,35 @@
 
   float: right;
 }
+
+.table {
+  border-collapse: collapse;
+  width: 100%;
+}
+
+.table td, .table th{
+  border: none;
+  padding: 8px;
+}
+
+.table tr:nth-child(odd) {
+  background-color: #f4f5f5;
+}
+
+.table tr:nth-child(even) {
+  background-color: white;
+}
+
+.table tr:hover {background-color: #ddd;}
+
+.table th {
+  padding-top: 12px;
+  padding-bottom: 12px;
+  text-align: left;
+  background-color:#f4f5f5;
+  color: rgba(36, 41, 46, 0.75);
+}
+
+.table td:first-child {
+  width: 9%;
+}
