+++ web/ui/src/features/component/ComponentList.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
 
-import { HealthLabel } from '../component/HealthLabel';
-import { ComponentInfo } from '../component/types';
-import Table from '../widget/Table';
+import { HealthLabel } from './HealthLabel';
+import { ComponentInfo } from './types';
 
 import styles from './ComponentList.module.css';
 
@@ -11,33 +10,30 @@ interface ComponentListProps {
   components: ComponentInfo[];
 }
 
-const TABLEHEADERS = ['Health', 'ID'];
-
 const ComponentList = ({ components }: ComponentListProps) => {
-  const tableStyles = { width: '100px' };
-
-  /**
-   * Custom renderer for table data
-   */
-  const renderTableData = () => {
-    return components.map(({ health, id }) => (
-      <tr key={id}>
-        <td>
-          <HealthLabel health={health.state} />
-        </td>
-        <td>
-          <span>{id}</span>
-          <NavLink to={'/component/' + id} className={styles.viewButton}>
-            View
-          </NavLink>
-        </td>
-      </tr>
-    ));
-  };
-
   return (
     <div className={styles.list}>
-      <Table tableHeaders={TABLEHEADERS} renderTableData={renderTableData} style={tableStyles} />
+      <table className={styles.table}>
+        <tr>
+          <th>Health</th>
+          <th>ID</th>
+        </tr>
+        {components.map((component) => {
+          return (
+            <tr>
+              <td>
+                <HealthLabel health={component.health.state} />
+              </td>
+              <td>
+                {component.id}
+                <NavLink to={'/component/' + component.id} className={styles.viewButton}>
+                  View
+                </NavLink>
+              </td>
+            </tr>
+          );
+        })}
+      </table>
     </div>
   );
 };
