+++ web/ui/src/features/component/ComponentView.tsx
@@ -3,12 +3,12 @@ import { Link } from 'react-router-dom';
 import { faCubes, faLink } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
-import { partitionBody } from '../../utils/partition';
+import { RiverValue } from '../../features/river-js/RiverValue';
+import { AttrStmt, Body, StmtType } from '../../features/river-js/types';
 
-import ComponentBody from './ComponentBody';
 import ComponentList from './ComponentList';
 import { HealthLabel } from './HealthLabel';
-import { ComponentDetail, ComponentInfo, PartitionedBody } from './types';
+import { ComponentDetail, ComponentInfo } from './types';
 
 import styles from './ComponentView.module.css';
 
@@ -128,5 +128,121 @@ export const ComponentView: FC<ComponentViewProps> = (props) => {
         )}
       </main>
     </div>
+  );
+};
+
+/**
+ * partitionBody groups a body by attributes and inner blocks, assigning unique
+ * keys for each.
+ */
+function partitionBody(body: Body, rootKey: string): PartitionedBody {
+  function impl(body: Body, displayName: string[], keyPath: string[]): PartitionedBody {
+    const attrs: AttrStmt[] = [];
+    const inner: PartitionedBody[] = [];
+
+    const blocksWithName: Record<string, number> = {};
+
+    body.forEach((stmt) => {
+      switch (stmt.type) {
+        case StmtType.ATTR:
+          attrs.push(stmt);
+          break;
+        case StmtType.BLOCK:
+          const blockName = stmt.label ? `${stmt.name}.${stmt.label}` : stmt.name;
+
+          // Keep track of how many blocks have this name so they can be given unique IDs.
+          if (blocksWithName[blockName] === undefined) {
+            blocksWithName[blockName] = 0;
+          }
+          const number = blocksWithName[blockName];
+          blocksWithName[blockName]++;
+
+          const key = blockName + `_${number}`;
+
+          inner.push(impl(stmt.body, displayName.concat([blockName]), keyPath.concat([key])));
+          break;
+      }
+    });
+
+    return {
+      displayName: displayName,
+      key: keyPath,
+      attrs: attrs,
+      inner: inner,
+    };
+  }
+
+  return impl(body, [rootKey], [rootKey]);
+}
+
+interface PartitionedBody {
+  /** key is a list of unique identifiers for this partitioned body. */
+  key: string[];
+  /** displayName is a list of friendly identifiers for this partitioned body. */
+  displayName: string[];
+
+  attrs: AttrStmt[];
+  inner: PartitionedBody[];
+}
+
+interface ComponentBodyProps {
+  partition: PartitionedBody;
+}
+
+const ComponentBody: FC<ComponentBodyProps> = ({ partition }) => {
+  const sectionClass = partition.key.length === 1 ? '' : styles.nested;
+
+  return (
+    <>
+      <section id={partition.key.join('-')} className={sectionClass}>
+        {
+          // If the partition only has 1 key, then make it an h2.
+          // Otherwise, make it an h3.
+          partition.displayName.length === 1 ? (
+            <h2>{partition.displayName}</h2>
+          ) : (
+            <h3>
+              {partition.displayName.map((val, idx) => {
+                return (
+                  <Fragment key={idx.toString()}>
+                    <span>{val}</span>
+                    {idx + 1 < partition.key.length && <span> / </span>}
+                  </Fragment>
+                );
+              })}
+            </h3>
+          )
+        }
+        <div className={styles.sectionContent}>
+          {partition.attrs.length === 0 ? (
+            <em className={styles.informative}>(No set attributes in this block)</em>
+          ) : (
+            <table>
+              <thead>
+                <tr>
+                  <th className={styles.nameColumn}>Name</th>
+                  <th className={styles.valueColumn}>Value</th>
+                </tr>
+              </thead>
+              <tbody>
+                {partition.attrs.map((attr) => {
+                  return (
+                    <tr key={attr.name}>
+                      <td className={styles.nameColumn}>{attr.name}</td>
+                      <td className={styles.valueColumn}>
+                        <RiverValue value={attr.value} />
+                      </td>
+                    </tr>
+                  );
+                })}
+              </tbody>
+            </table>
+          )}
+        </div>
+      </section>
+      {partition.inner.map((body) => {
+        return <ComponentBody key={body.key.join('.')} partition={body} />;
+      })}
+    </>
   );
 };
