+++ web/ui/src/features/river-js/RiverValue.tsx
@@ -1,4 +1,4 @@
-import React, { FC, Fragment, ReactElement } from 'react';
+import { FC, Fragment, ReactElement } from 'react';
 
 import { ObjectField, Value, ValueType } from './types';
