--- web/ui/src/features/widget/Table.module.css
@@ -1,37 +1,0 @@
-.table {
-  border-collapse: collapse;
-  width: 100%;
-  table-layout: fixed;
-}
-
-.table td, .table th{
-  border: none;
-  padding: 8px;
-}
-
-.table tr:nth-child(odd) {
-  background-color: #f4f5f5;
-}
-
-.table tr:nth-child(even) {
-  background-color: white;
-}
-
-.table tr:hover {background-color: #ddd;}
-
-.table th {
-  padding-top: 12px;
-  padding-bottom: 12px;
-  text-align: left;
-  background-color:#f4f5f5;
-  color: rgba(36, 41, 46, 0.75);
-}
-
-.table td:first-child {
-  width: 2%;
-}
-
-.table span {
-  white-space: normal;
-  word-break: break-all;
-}
